Do not return glGetString error on desktop GL context
Nvidia shield devices support using egl to switch to a full desktop
opengl context. In opengl 3.0+, GL_EXTENSIONS have to be retrieved
with glGetStringi and is invalid for glGetString. Thus,
eglMakecurrent will return a gl error if the case is not handled.
The logic here is building a wrapper for glGetStringi, thus the
error can be ignored. This patch allows GL_EXT_debug_marker to
still pushed into the tokenized extension string used by the
glGetStringi wrapper.

Change-Id: I9c599e10c62aabf684bde4e81719aa248327ac80
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>